refactor: split cli command parsing tests
